Synchronizing Password and Key dialogs with UI
Added some logic to get the ssh key and password dialogs running on the
UI thread when they are not already.

This works for both my viz service branch and the master branch where
*everything* is done on the UI thread.

I also changed the OK button on the password dialog to not have the
border, which doesn't match standard looking widgets.